job-manager: track pending allocs in one place
Problem: in limited mode, scheduler alloc requests are kept in
a list, but in unlimited mode they are tracked with a counter,
which complicates the code for likely little performance gain.

Eliminate the counter and always use the list.  Since deletion from
the list is idempotent this will probably be more reliable.