job-manager: don't store ephemeral events
Problem: events posted with the EVENT_NO_COMMIT flag are
kept in job->eventlog, contributing to memory growth.

These events are not important and need not be kept.  This does mean
that journal consumers like job-list will not see those events when they
are reloaded and the journal is replayed, but since the idea is that
these events are ephemeral, that is probably OK.  Nothing is lost when
the instance restarts since these events were never stored in the KVS.

Fix up one test that was expecting job list to recreate annotations
after a reload.