job-manager: add journal event count to stats
Problem: when the scheduler was malfunctioning and spamming
the job manager with annotations, we had no tools for diagnosis.

Count events that are posted to the journal and include this count
in the 'flux module stats job-manager' output.  A rapidly increasing
event count would have been a helpful clue in the above case.