libschedutil: avoid multiple free on error path
Problem: when handling a successful alloc callback from the scheduler,
if the KVS commit of R fails (unlikely), the rank 0 broker may crash
with a double free.

In the KVS commit continuation, take the future off of the global cleanup
list before checking for KVS error, not after, because the future is
destroyed on all this function's return paths.

Since the alloc context is in the future aux container with alloc_destroy()
registered as a destructor, and the future is destroyed on the function's
error path, drop the additonal call to alloc_destroy() on the error path.