job-exec: add a start barrier timer
Problem: A common failure mode on large systems is that one or
more job shells hang during initialization, typically when trying
to change the working directory to a parallel filesystem that is
having problems.  In this case, the job hangs before the `start`
event until timeout. It would be preferrable to have some sort of
timeout to catch this condition earlier.

Add a timeout to the initial exec shell barrier. The timer starts
when the first shell enters the barrier. If the timeout elapses, an
exception is raised on the job and the ranks that have not checked
in are drained with a message indicating a possibe node hang.

The timeout is 30m by default, with an ability to override for now via
a jobspec attribute attributes.system.exec.barrier-timeout (floating
point seconds).

The timeout may be disabled by setting barrier-timeout=0.