job-exec: fix potential job reference leak
Problem: If the rank 0 job shell is killed or exits when the job-exec
module sends a lost shell RPC to it, then there may never be a
response and the future from that RPC will be leaked. Additionally,
jobinfo_shell_rpc_pack() takes a reference on the jobinfo object,
which is only dropped when the future is destroyed. Therefore, this
leaves the jobinfo object for a complete job active until the job-exec
module is unloaded.

Do not wait for the response from jobinfo_shell_rpc_pack(), instead
destroy the future immediately.