job-manager: move bad jobs to KVS lost+found
Problem: when KVS job entries cannot be replayed by the job manager,
they are ignored by flux but left in the KVS.  A purge will not touch
them since that is based on the in-memory inactive job hash.

Move bad job directories to a KVS lost+found.jobs directory for manual
cleanup/analysis.

Fixes #6133