job-exec: return error if bulkexec jobspec config object unpack fails
Problem: Errors from parsing the attributes.system.exec.bulkexec object
in jobspec are ignored. While this is mainly used in testing, a silent
failure here could result in invalid or mysteriously failing tests.

Switch to `json_unpack_ex()`, make all keys optional, and add a `!`
at the end of the bulkexec object so that extra keys raise an error.

On error, return the error.text in a new `flux_error_t` parameter
to exec_ctx_init().