testsuite: fix mock_exception argument in bulksubmit tests
Problem: A test in t2713-python-cli-bulksubmit.t sets
attributes.system.exec.bulkexec.mock_exception=0 to disable the
mock exec exception, but this relies on an ignored json_unpack()
error since the expected type of the mock_exception key is string.

Use "none" instead of 0 to set but disable the mock exception so this
doesn't break when errors are checked on unpack of the bulkexec object.