testsuite: add flux-resource status `torpid` tests
Problem: No tests in the testsuite check the handling of the torpid
virtual state in `flux resource list`.

Add a test input object to t/flux-resource/status/torpid.json
with some torpid ranks that overlap drain ranks, along with
the expected output in torpid.expected. The loop over inputs in
t2351-resource-status-input.t then serves to check that torpid ranks
are displayed, and that they do not include already drained ranks.

Then, add a specific test at the end of t2351 that ensures all torpid
ranks are included in output when the drained/draining states are
not included in output.