python: accept multiple queues in job_list(), JobList
Problem: There is no way to easily get a list of jobs for
multiple queues at once with the Python flux.job.job_list() and
flux.job.JobList.

If the queue parameter to `job_list()` is iterable and not a string,
pass it along to the queue constraint as a list, which is the same
as an "or" between all queue elements. Since JobList calls job_list()
it also supports a list of queues.