python: support multiple queues in JobStats
Problem: The JobStats class only supports a single queue.

Support a list (or other iterable) for the queues parameter of the
JobStats class.

Move the list of stats to a set of class variables to avoid repetition.

Add an internal QueueStats class to hold specific queue stats, and
store any queue-specific stats in a dictionary of these objects by
queue name.

If multiple queues are specified in the JobStats constructor, just
sum the individual queue stats together before creating any derived
statistics.