coverage: atomically update gcov files
problem: we keep getting merge issues with gcov files.  It seems the
default for updating these is "single" meaning single threaded unless
the compilation is done with `-pthread`. I'm guessing we see the errors
most on libraries we use in multiple places that are not themselves
compiled that way, even though they're eventually linked into binaries
that are. If this doesn't cover it, we should also do what's listed in

solution: add `-fprofile-update=atomic` to our coverage flags