flux-resource: use a set type for args.queue
Problem: The `-q, --queue` option for all flux-resource(1) subcommands
takes a comma separated list of queues and args.queue is a string
which has to be split at the comma in place when used. This not only
duplicates work, but is error prone since the string itself is also
iterable and a forgotten args.queue.split() would cause the code to
treat each character as a queue.

Store args.queue as a set() using flux.util.FilterActionSetUpdate.
Update sites where args.queue.split(",") is no longer necessary.