flux-resource: add queue argument to ResourceSetExtra initializer
Problem: An empty resource set cannot have any associated properties
and therefore cannot be in a queue. However, in `flux resource list`
output, it may be useful to display an "empty" line of output for one
or more queues. For instance, when no nodes are allocated, down, etc.

Allow the ResourceSetExtra class to have a forced queue associated
with it. If set in the initializer, then instead of iterating
properties to deterimine the resource set queue the manually set
queue is simply returned. This allows an empty resource set to have
an associated queue.