flux-resource: ensure empty resource sets can display a queue
Problem: When there are no resources in a given state, flux-resource(1)
list creates an empty resource set for that state so that a line with
zero resources appears in output. However, this is confusing when there
are configured queues because the {queue} column will be empty since
an empty resource set cannot have any associated properties or queues.

Now that the ResourceSetExtra class allows a queue to be set manually,
create an empty resource set for each configured queue (or if -q,
--queue was specified, each queue given on the cmdline). This will
cause an empty resource set for each queue to be displayed in output.

Fixes #6166