testsuite: remove test for empty queue in t2350-resource-list.t
Problem: In the 'flux resource lists expected queues...' tests in
t2350-resource-list.t, the test ensures no queue names are listed in
the output of empty resource sets (i.e. where nnodes=0). However, this
is a bug or limitation, not a feature, so it should not be tested for.

Remove these tests.