testsuite: fix racy test in t2406-job-exec-cleanup.t
Problem: The kill-timeout test races with max-kill-count in
t2406-job-exec-cleanup.t. That is, in some cases, max-kill-count
(which is forced to 3 attempts in the test) is triggered before the
test checks that the kill-timeout value has been increased by the
job-exec module, leading to a spurious test failure.

Split up the tests for kill-timeout and max-kill-count. When testing
kill-timeout, set max-kill-count to a large number so it is unlikely
to be reached. Then separately test max-kill-count after reloading
the job-exec module with max-kill-count=3.

Fixes #6184