matrix/actions: add per-target timeout-minutes
problem: we have widely varying timouts based on the target and the
usage, and only one timeout value

solution: add timeout-minutes to the matrix and expand it on the
docker-run-checks step.  The downside is that this will not time out if
another step stalls, but which other step would stall?