content-sqlite: allow sqlite configs to be mutable
Problem: Internally the configs for journal_mode and synchronous
sqlite settings are stored as "const char *".  This makes the
configs immutable.  In the near future they could change, such as
if config file settings override the variables.

Instead of storing sqlite configs as "const char *" types, change
types to "char *".  Use strdup() to store the configuration
setting, so that it can be freed/changed at a later time.