content-sqlite: split truncate out of function
Problem: In the near future we may wish to open the sqlite multiple
times within the content-sqlite module.  The current
content_sqlite_opendb() takes a "truncate" parameter that would truncate
the db on every open.  This is not what we want if we are
closing/opening the database multiple times.

Solution: Split out the truncate into a new "setup" function.  This
setup function will be called once when the module is loaded.