content-sqlite: support preallocate config
Problem: When a disk runs out of space, the content-sqlite module
can no longer work.  It would be convenient if space could be
reserved ahead of time to prevent these types of problems.

Support a new preallocate module and config option that takes a byte maximum
as input.  This option will internally create a new special database and
write data to that database until the byte max is reached.
Then this database will be dropped.  This internally reserves space for the sole
use of sqlite.

Note that this feature will not work if any type of journaling or
write ahead log is used, as that will also require disk space.