flux-jobs: add -i, --include=HOSTS|RANKS option
Problem: There is no easy way to restrict the output of flux-jobs(1)
based on a set of hosts or ranks, but this is basic and useful
functionality.

Add a new `-i, --include=` optiont that takes a hostlist or idset of
targets, and adds a constraint to the JobList query that restricts
matching jobs to those that were allocated the provided targets.

Fixes #6202