kvs: correct code logic about what is an append
Problem: In PR #2547, a fix was added into the KVS to deal with duplicate
appends to a KVS entry.  However, a corner case was left in the code.
When an append is done to a KVS entry that does not exist, the append
is treated like a normal insertion.  This corner case should also be
counted as an "append".  Otherwise, duplicate appends could happen at
the beginning of a KVS entry when it is created.

Solution: Set the append flag to true when a KVS entry does not exist
and the append is treated like an insert.

Fixes #6207