content-sqlite: disable journaling if hit ENOSPC
Problem: If a disk fills up, sqlite may no longer be able to operate
because journal files can no longer be written to disk.  However,
if space was pre-allocated, sqlite can still be used if we turn off
journaling.

If ENOSPC is hit and pre-allocated space was configured, turn off
journaling in an attempt to keep the content-sqlite module functional,
although with slower performance.