perilog: switch plugin to bulk-exec
Problem: The perilog plugin can only execute a single prolog or
epilog task on rank 0, but it may be useful and more efficient to
allow prolog/epilog tasks across multiple ranks in the future.

Update the perilog plugin to use the libsubprocess/bulk-exec API.
Keep the current behavior of executing a single process on rank 0.
There should be no functional difference with this change.