cron: do not force a working directory in cron tasks
Problem: When cron tasks are created without an explicit working
directory, the current working directory of the broker is used in the
flux_cmd_t. However, this is unnecessary and forces the cron task to
fallback from using posix_spawn to fork/exec, which is a performance
issue, especially when the broker RSS is large.

Allow cron tasks to have an "unset" (NULL) cwd. This avoids the
fallback to fork/exec while still keeping the cwd the same as the
broker.

Fixes #6104