perilog: do not raise a job exception when prolog is canceled
Problem: When the prolog is canceled due a job exception while the
prolog is running (e.g. user pressed Ctrl-C), the prolog fails due
to the signal and another job exception is raised because of the
prolog failure. This is unnecessary.

Skip the second, unnecesary job exception when the prolog has been
canceled.