sched-simple: abort scheduler if free request fails
Problem: Free requests to the scheduler can fail without detection
during the testsuite because the scheduler logs an error, but no
other error is raised.

Abort the scheduler (stop the reactor with an error) when a free
request fails so that the testsuite fails on this condition.

Note: we can't raise a job exception because the job causing the error
has likely proceeded to inactive by this time, and since sched.free
has no response, the best solution here is a fatal error.