libev: fix memory fence on i386 under valgrind
problem: libev accesses one byte past the end of the stack in its custom
assembly memory fence.  I have no idea why, it doesn't help, and causes
valgrind to throw a warning every single time it gets invoked.  @garlick
investigated this previously in #1897 and suppressed it in core, but
that never made it to sched.  In investigating I found a patch for it
that has the exact same effect but on a valid location.

solution: Apply this patch identified for debian upstream in 2017 so it
will stop doing that: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=850861