perilog: add plugin.query callback
Problem: The perilog plugin doesn't offer a way to query its current
configuration or the currently executing set of prolog/epilogs and
which the ranks on which they are still active.

Add a plugin.query callback to the plugin which reports the current
prolog and epilog configuration (if any), and all currently executing
perilog "processes" by jobid, including the ranks on which they are
still active.