python: let `--` separate non-option args in submission cli commands
Problem: Python argparse apparently attempts to process arguments past
the first unknown arg, even when using parse_known_args(). This doesn't
seem to cause a problem unless the non-option argument happens to be
ambigious given all known args, then Python will throw an exception
immediately, even though the option argument is valid for the user's
command/script etc. For example:

 $ flux submit -n2 myapp --n=2
 usage: flux submit [OPTIONS...] COMMAND [ARGS...]
 flux submit: error: ambiguous option: --n=2 could match --nodes, --ntasks

A workaround for this problem has not been found yet, possibly the
issue will be or is resolved in future versions of Python.

One solution is to separate flux run/submit/batch/alloc arguments
from the command/script and its arguments with the standard `--`,
but Flux currently then thinks this "argument" is part of the command.

In order to allow `--` to be used to workaround this Python issue,
remove it from args.command (for flux run/submit) and args.SCRIPT
(for flux batch). It is not necesssary to explicitly ignore `--`
with `flux alloc`, since the `--` is passed through to `flux broker`,
which uses GNU getopt and therefore ignores `--` by default.

Fixes #6228