perilog: improve error message on failure
Problem: The error message logged by the perilog plugin for failed
prolog/epilog is not useful in its current form now that the plugin
uses bulk-exec to potentially launch multiple processes.

Simplify the log message to duplicate what is done in the housekeeping
service of the job manager. Include the rank, job id, and a textual
error message instead of a useless exit code.