perilog: allow epilog to be canceled on exception for testing
Problem: It would be convenient for testing purposes if the job epilog
could be canceled on exception, but currently only the prolog can be
canceled in this way.

Support canceling the epilog via a new config key cancel-on-exception.
If unset, then it defaults to true for the prolog and false for
the epilog.

This commit also cleans up handling of the kill_timeout by moving
it inside the perilog_proc struct since it no longer applies to only
the prolog.