perilog: support "inf", "infinity" for prolog/epilog timeout
Problem: Since fsd_parse_duration() is used to parse the perilog
timeout configuration, "inf" and "infinity" are supported. This would
cause a value of INFINITY to be passed to the flux timer watcher,
which doesn't seem useful. Also, this value can't be encoded by JSON,
which would cause the plugin query callback to fail.

Just make "infinity" an alias for "0", i.e. the timeout is disabled.