flux-hostlist: add a timeout for reading from stdin
Problem: A recent update to flux-hostlist(1) changed the behavior when
the command is called with no arguments. This could lead to a hang
if the command was previously used to gather the hostlist for a job,
for example.

As a mitigation, add a 15s timeout for reading from stdin. For
testing purposes, allow the timeout to modified wtih a
FLUX_HOSTLIST_STDIN_TIMEOUT environment variable.