perilog: start epilog even if prolog fails or is canceled
Problem: When the prolog fails for any reason, including a timeout,
cancelation, or a failure on one or more nodes, the associated job
is not started and therefore any configured epilog is not started.
This is problematic if the prolog did something that is assumed to
be undone by the epilog.

Start any configured epilog immediately when a prolog fails.
The epilog-start event is posted before the prolog-finish in order
to prevent the job from reaching INACTIVE state immediately.