resource: add resource.scheduling config key for amending R
Problem: When the resources for an instance are configured using the
resource.config array, there is no way to amend the result with a
scheduling key, which may be required on some systems.

Add a new resource.scheduling config key which contains the path to
a JSON object that will be loaded and used to amend the configured R.

This allows the scheduling JGF to be kept separately from the simpler
R configuration, whether using resource.path or resource.config.