flux-resource: support `allocated` state in `flux resource status`
Problem: The `flux resource status` command does not support
displaying nodes in the `allocated` state, even though the underlying
`ResourceStatus` object used in the command has this information.

Add `allocated` to the list of valid states. As a side-effect, the
set of allocated nodes will be displayed with `flux resource status
-s all`.