validator: add configurable job size for require-instance
Problem: The require-instance validator plugin is an all-or-nothing
proposition -- either all jobs must be instances of Flux or none are
required to be. This is limiting, since there are times when admins
may only want to limit large jobs to being new instances of Flux to
lessen the impact on a system instance.

Add two options to the require-instance validator to set a minimum
size for instances to which the requirement applies. The default is
0 so that the plugin applies to all instance by default, but with
--require-instance-minnodes=N and --require-instance-mincores=N this
default can be increased so that only jobs above this size will be
rejected if they are not new instances of Flux.

To avoid any issue when minnodes is set but not minncores, set
conservative default cores-per-node of 16 when mincores is not
set. This will likely limit cores-only jobs more than jobs that set
a node count, but this should encourage admins to set both, or users
to use -N, --nodes with their jobs.