flux-resource: include housekeeping nodes in output
Problem: The set of ranks currently engaged in housekeeping is not
included in the output of `flux resource status`, but this would be
a simple way to make users and admins aware that some resources are
busy doing housekeeping, so it should be included.

Add support for a housekeeping state in `flux resource status` output.
Add housekeeping to the default list of states to display (if no ranks
are in housekeeping then this line is suppressed by default). Increase
the width for the {state} column since housekeeping has more than
10 characters.