python: add conf_get() method to Flux handle
Problem: There is no convenient method to fetch and query the Flux
configuration from Python.

Add a new method to the flux.Flux handle object, conf_get().

On the first use, or if the optional `update` argument is True,
this method synchronously fetches the current config from the broker
and caches it for future use.

The function takes an optional `key` parameter, which is provided
as a dotted-string form, e.g. `tbon.topo` to return a specific value
from the current config. If the value does not exist, a default value
is returned.