shell: set output limit for single user instances
Problem: The KVS has a size limit of INT_MAX for when returning kvs
values.  This limit can be exceeded by a job's standard output because
it is continually appended and the total size is not yet tracked by
the KVS.  When reading the output later, such as via `flux job attach`,
this can lead to EOVERFLOW errors.

Solution: For a single user instance, default to a maximum standard
output of 1G instead of "unlimited".  1G should provide a practical
maximum for most users and encourage them to send standard output to
a file if they want to save excess standard output.

Do not allow configuration larger than this.  As a consequence the
configuration of "unlimited" is no longer allowed.

Fixes #6256