python: add flux handle attribute to ValidatorPlugin base class
Problem: Validator plugins access a per-thread on-deman Flux handle
via the ValidatorJobInfo class. But objects of this class are not
available in the `configure` method, and also this means that a
handle may be created and destroyed for each job being validated,
instead of persisting for the lifetime of the plugin thread.

Add an on-demand handle in th ValidatorPlugin base class instead.
This will allow the plugin to access a Flux handle at any time,
as well as persisting for the lifetime of the plugin. Finally,
this handle doesn't have to be installed in thread-local storage,
since each plugin is run in a separate thread as it is.