ipaddr: allow PMI interface to be selected by CIDR
Problem: interface addresses are not always predictable, so an
alternate method of selecting an IP interface for Flux to use when
bootstrapping via PMI is needed.

Allow FLUX_IPADDR_INTERFACE to also be an IPv4 network address in
CIDR format, e.g. "192.168.1.0/24".

This commit does not add support for IPv6 CIDR.