broker: add tbon.interface-hint
Problem: the selection of the bind address is only influenced
with environment variables.

FLUX_IPADDR_INTERFACE and FLUX_IPADDR_HOSTNAME are the env vars.
Add a tbon.interface-hint configuration key and broker attribute
that perform the same function.

By default the value is "default-route".
Setting it to "hostname" is equivalent to setting FLUX_IPADDR_HOSTNAME.
Otherwise it is equivalent to setting FLUX_IPADDR_INTERFACE.

Leave the env vars as is for now since they are the only way to globally
influence all Flux instances on a system.

Fixes #5754