flux-start: implement PMI abort callback
Problem: if the flux broker calls the PMI abort function,
flux-start (the PMI server) is not notified.

Add an abort callback that logs a message and asks all subprocesses
to terminate immediately.

Update tbon.interface-hint test that was expecting broker to exit 1
on a bad hint.  The broker is now terminated with SIGKILL.

Update another tbon.interface-hint test that required a ratcheted down
timeout.  That test now fails immediately.