python: support auto-expanding fields in OutputFormat
Problem: When generating formatted output, there are times where
truncating a wide field is not desirable, for example when a comma
separated list is being displayed, truncation can completely hide
some entries. However, the only current alternative is to use a
very wide field in the output format by default, which is also not
desirable since it wastes horizontal space to handle an uncommon case.

Extend the current sentinel `?:`, which specifies that a field is
to be collapsed if the output width for all entries is 0, with a
new sentinel `+:`, which indicates the field width should be increased
to handle the largest width of all displayed items. To take both
actions `?+:` is also supported.

Extend `OutputFormat.filter_empty()` to handle these new sentinels
and modify the format spec if necessary before returning the new
format to the caller.