broker: set tbon.interface-hint from enclosing instance
Problem: we don't have a way to configure subinstances to
use a particular network, other than setting the global
environment variable FLUX_IPADDR_INTERFACE.

Change the way the default tbon.interface-hint is set.  Instead
of hard coding "default-route", use the value that is set in the
enclosing instance's PMI KVS, if any.

This would allow a system instance to configure tbon.interface-hint
and batch/alloc jobs that it starts would use that value unless
overridden.  If the value is not configured in the system instance,
then the attribute will have the default value which will be inherited
by the child.

Fixes #6272