flux-resource: use expandable formats in `flux resource list`
Problem: The fixed width of the queue field in `flux resource list`
causes confusion for users with overlapping queues because the queues
are joined into a comma separated list and truncation could cause
some queues to be completely left out of the field.

Now that the OutputFormat class supports "expandable" fields, use this
with the `queue` field in `flux resource list` so that no truncation
is guaranteed without using a large width by default.

Also apply the expandable field to ngpus to save some horizontal
space in the common case.