libsubprocess: output correct service name
Problem: The service name used when setting up a subprocess server
is configurable.  However, most of the error messages assume the
service name is "rexec".  This can lead to log messages pointing
users to the wrong RPC target.

Store the service name in the subprocess server context and output
the appropriate RPC target in error messages.