broker/boot_pmi: refactor instance-level setting
Problem: adding support for reading flux.tbon-interface-hint from
the PMI KVS is awkward due to the structure of an internal function.

Refactor set_instance_level_attr() so that it returns a boolean
'under_flux' flag indicating whether this instance was launched by Flux.
Use that that flag to determine whether to set the jobid attr.
The flag will also be useful for gating whether to check for the
interface hint in the PMI KVS.