libsubprocess: support fbuf credit callback
Problem: It would be convenient to be notified when data has
been read from a buffer, so that more can be written into it.

Support a credit callback.  The callback will be called after some
data has been read out of the buffer and inform the caller of
the number of bytes read.

Add unit tests.