resource: add mode=force to undrain RPC
Problem: The resource module always aborts an entire undrain RPC if
any target is not currently drained, but this makes it awkward to
force the undrain of a set of targets regardless of whether they are
currently drained.

Add an optional "mode" to the undrain RPC payload. If set with a
value of "force", remove currently undrained ranks from the set
of provided targets instead of aborting the entire operation and
returning an error.