perilog: fix kill of prolog/epilog when using IMP
Problem: When using the IMP to run the prolog/epilog in the perilog
plugin, the IMP path is not set in the bulk_exec object, so it doesn't
know to use `flux-imp kill` to signal the subprocess. This results in
failure to cancel job prologs when a job gets an exception.

Add a boolean to the perilog configuration if using the IMP. When
this boolean is true, set the IMP path in the bulk_exec implementation
so that it uses `flux-imp kill`.

Fixes #6296.