perilog: add remaining time to plugin query response
Problem: There's no way to easily see much time a prolog/epilog with
a timeout has remaining.

Add a remaining_time entry to the `flux jobtap query` response for the
perilog plugin for each perilog proc with a timeout. If the prolog or
epilog does not have a configured timeout, then -1 will be displayed.