perilog: add call to optionally drain ranks, post finish event
Problem: In the perilog plugin, the code to drain ranks and post the
prolog/epilog-finish event is embedded in the bulk-exec completion
callback, making it difficult to reuse outside that context.

Abstract that functionality into a proc_drain_and_finish() function,
which optionally drains ranks that failed or are still active, then
posts the appropriate finish event.