job-info: distinguish bad job ID from missing key
Problem: when flux job info fails, it does not print a different
error message for a bad job ID vs missing key.

Ensure that when the job owner is not found in the LRU cache,
the eventlog is looked up, even for the instance owner.
Then if the lookup of the eventlog fails with ENOENT, we know
it's a bad job ID.

Fixes #6325