doc: add warning about stdio file paths in submission cli docs
Problem: It may surprise users that --output, --error, and --input
files are opened on the first node of the allocation and relative
to the working directory of the job, not on the submission node.

Add a note to the common job-standard-io.rst document warning users
of this potential pitfall.

Fixes #6329