testsuite: generalize error regex in t2311-resource-drain.t
Problem: A couple tests in t2311-resource-drain.t check for a specific
error message when ranks are not drained with flux-resource undrain,
but this doesn't allow for future improvement of the message.

Check for a pattern instead of an exact string so that future improved
error messages will also match.