resource: include hostnames in undrain error response
Problem: The undrain error response when encountering a target that
is not currently drained includes only the first affected rank.

Iterate over all targets and build a list of ranks that are not
currently drained. Convert the ranks to a set of hostnames to include
in the error response.

Fixes #6319